Move multisig code out of main
